requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
Django==4.0
django-cors-headers==4.4.0
django-filter==2.4.0
djangorestframework==3.15.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
packaging==24.0
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.8.0
PyMySQL==1.1.0
pypng==0.20220715.0
python-weixin==0.5.0
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
webencodings==0.5.1
xmltodict==0.13.0
